refactor: components 폴더 구조 변경
DrinkSection과 MoneySection으로 분리
